+++ packages/gestalt-codemods/1.8.0-1.9.0/__tests__/remove-responsive-text-sizing.test.js
@@ -1,0 +1,19 @@
+import { defineTest } from 'jscodeshift/dist/testUtils.js';
+
+jest.mock('../remove-responsive-text-sizing', () => {
+  return Object.assign(
+    require.requireActual('../remove-responsive-text-sizing'),
+    { parser: 'flow' }
+  );
+});
+
+describe('remove-responsive-text-sizing', () => {
+  ['remove-responsive-text-sizing-transform'].forEach(test => {
+    defineTest(
+      __dirname,
+      'remove-responsive-text-sizing',
+      { quote: 'single' },
+      test
+    );
+  });
+});